+++ src/JiraStatusNotifier/Channel/ChannelInterface.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Chemaclass\JiraStatusNotifier\Channel;
 
 use Chemaclass\JiraStatusNotifier\Jira\ReadModel\Company;
+use Chemaclass\JiraStatusNotifier\Jira\TicketsByAssignee\TicketsByAssignee;
 
 interface ChannelInterface
 {
