+++ src/JiraStatusNotifier/Channel/Cli/Channel.php
@@ -7,8 +7,8 @@ namespace Chemaclass\JiraStatusNotifier\Channel\Cli;
 use Chemaclass\JiraStatusNotifier\Channel\ChannelInterface;
 use Chemaclass\JiraStatusNotifier\Channel\ChannelResult;
 use Chemaclass\JiraStatusNotifier\Channel\ReadModel\ChannelIssue;
-use Chemaclass\JiraStatusNotifier\Channel\TicketsByAssignee;
 use Chemaclass\JiraStatusNotifier\Jira\ReadModel\Company;
+use Chemaclass\JiraStatusNotifier\Jira\TicketsByAssignee\TicketsByAssignee;
 
 final class Channel implements ChannelInterface
 {
