+++ src/JiraStatusNotifier/Jira/TicketsByAssignee/FilterStrategy/FilterStrategy.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Chemaclass\JiraStatusNotifier\Jira\TicketsByAssignee\FilterStrategy;
+
+use Chemaclass\JiraStatusNotifier\Jira\ReadModel\JiraTicket;
+
+interface FilterStrategy
+{
+    public function isValidTicket(JiraTicket $ticket): bool;
+}
