+++ tests/JiraStatusNotifierTest/Unit/Jira/TicketsByAssignee/FilterStrategy/TicketFilterTest.php
@@ -2,28 +2,28 @@
 
 declare(strict_types=1);
 
-namespace Chemaclass\JiraStatusNotifierTests\Unit\IgnoreStrategy;
+namespace Chemaclass\JiraStatusNotifierTests\Unit\Jira\TicketsByAssignee\FilterStrategy;
 
-use Chemaclass\JiraStatusNotifier\IgnoreStrategy\TicketIgnorer;
 use Chemaclass\JiraStatusNotifier\Jira\ReadModel\Assignee;
 use Chemaclass\JiraStatusNotifier\Jira\ReadModel\JiraTicket;
 use Chemaclass\JiraStatusNotifier\Jira\ReadModel\TicketStatus;
+use Chemaclass\JiraStatusNotifier\Jira\TicketsByAssignee\FilterStrategy\TicketFilter;
 use DateTimeImmutable;
 use PHPUnit\Framework\TestCase;
 
-final class IgnoreStrategyTest extends TestCase
+final class TicketFilterTest extends TestCase
 {
     /** @test */
     public function shouldIgnore(): void
     {
-        $ignoreStrategy = TicketIgnorer::byAssigneeKey('key1', 'key2');
+        $ignoreStrategy = TicketFilter::notWithAssigneeKeys('key1', 'key2');
         self::assertTrue($ignoreStrategy->shouldIgnore($this->newTicket('key1')));
     }
 
     /** @test */
     public function shouldNotIgnore(): void
     {
-        $policy = TicketIgnorer::byAssigneeKey('key1', 'key2');
+        $policy = TicketFilter::notWithAssigneeKeys('key1', 'key2');
         self::assertFalse($policy->shouldIgnore($this->newTicket('key3')));
     }
